Update VspAttribution to VSAttribution
